+++ ShoppingListApi/Controllers/UserController.cs
@@ -12,7 +12,7 @@ using ShoppingListApi.Model.ReturnTypes;
 
 namespace ShoppingListApi.Controllers
 {
-    [Route("api/[controller]")]
+    [Route("[controller]")]
     [ApiController]
     public class UserController(
         IListUserService listUserService,